requirements.txt
fastapi
uvicorn[standard]
python-dotenv
gspread
google-auth
